Split tests by algorithm
