+++ tests/test_api_vit_haploid.py
@@ -1,0 +1,99 @@
+import pytest
+
+from . import lsbase
+import lshmm as ls
+import lshmm.core as core
+import lshmm.vit_haploid as vh
+
+
+class TestViterbiHaploid(lsbase.ViterbiAlgorithmBase):
+    def verify(self, ts, scale_mutation_rate, include_ancestors):
+        for n, m, H_vs, s, e_vs, r, mu in self.get_examples_pars(
+            ts,
+            ploidy=1,
+            scale_mutation_rate=scale_mutation_rate,
+            include_ancestors=include_ancestors,
+            include_extreme_rates=True,
+        ):
+            num_alleles = core.get_num_alleles(H_vs, s)
+            V_vs, P_vs, ll_vs = vh.forwards_viterbi_hap_lower_mem_rescaling(
+                n=n,
+                m=m,
+                H=H_vs,
+                s=s,
+                e=e_vs,
+                r=r,
+            )
+            path_vs = vh.backwards_viterbi_hap(m=m, V_last=V_vs, P=P_vs)
+            path, ll = ls.viterbi(
+                reference_panel=H_vs,
+                query=s,
+                num_alleles=num_alleles,
+                prob_recombination=r,
+                prob_mutation=mu,
+                scale_mutation_rate=scale_mutation_rate,
+            )
+            self.assertAllClose(ll_vs, ll)
+            self.assertAllClose(path_vs, path)
+
+    @pytest.mark.parametrize("scale_mutation_rate", [True, False])
+    @pytest.mark.parametrize("include_ancestors", [True, False])
+    def test_ts_simple_n10_no_recomb(self, scale_mutation_rate, include_ancestors):
+        ts = self.get_ts_simple_n10_no_recomb()
+        self.verify(
+            ts,
+            scale_mutation_rate=scale_mutation_rate,
+            include_ancestors=include_ancestors,
+        )
+
+    @pytest.mark.parametrize("scale_mutation_rate", [True, False])
+    @pytest.mark.parametrize("include_ancestors", [True, False])
+    def test_ts_simple_n6(self, scale_mutation_rate, include_ancestors):
+        ts = self.get_ts_simple_n6()
+        self.verify(
+            ts,
+            scale_mutation_rate=scale_mutation_rate,
+            include_ancestors=include_ancestors,
+        )
+
+    @pytest.mark.parametrize("scale_mutation_rate", [True, False])
+    @pytest.mark.parametrize("include_ancestors", [True, False])
+    def test_ts_simple_n8(self, scale_mutation_rate, include_ancestors):
+        ts = self.get_ts_simple_n8()
+        self.verify(
+            ts,
+            scale_mutation_rate=scale_mutation_rate,
+            include_ancestors=include_ancestors,
+        )
+
+    @pytest.mark.parametrize("scale_mutation_rate", [True, False])
+    @pytest.mark.parametrize("include_ancestors", [True, False])
+    def test_ts_simple_n8_high_recomb(self, scale_mutation_rate, include_ancestors):
+        ts = self.get_ts_simple_n8_high_recomb()
+        self.verify(
+            ts,
+            scale_mutation_rate=scale_mutation_rate,
+            include_ancestors=include_ancestors,
+        )
+
+    @pytest.mark.parametrize("scale_mutation_rate", [True, False])
+    @pytest.mark.parametrize("include_ancestors", [True, False])
+    def test_ts_simple_n16(self, scale_mutation_rate, include_ancestors):
+        ts = self.get_ts_simple_n16()
+        self.verify(
+            ts,
+            scale_mutation_rate=scale_mutation_rate,
+            include_ancestors=include_ancestors,
+        )
+
+    @pytest.mark.parametrize("scale_mutation_rate", [True, False])
+    @pytest.mark.parametrize("include_ancestors", [True, False])
+    def test_ts_larger(self, scale_mutation_rate, include_ancestors):
+        ts = self.get_ts_custom_pars(
+            ref_panel_size=46, length=1e5, mean_r=1e-5, mean_mu=1e-5
+        )
+        self.verify(
+            ts,
+            scale_mutation_rate=scale_mutation_rate,
+            include_ancestors=include_ancestors,
+        )
